requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
psycopg[binary]>=3.2
python-dotenv>=1.0
httpx>=0.27,<0.28
python-multipart>=0.0.5
psycopg[binary]>=3.1,<4
